--- .firebase/hosting.YnVpbGQ.cache
@@ -1,16 +1,0 @@
-asset-manifest.json,1690486789161,ddfa416523b0676c583a1d69d179dba1f517b8b7bfc2db4a04cf21358253c714
-index.html,1690486789161,20f19caf002e02d3ef7da26328ffd90131438011db51794f241576bbfba18584
-manifest.json,1689833241906,5c997de1364b8be939319fa9209abd77f2caf7f8844999a9e2e9173f844e7840
-robots.txt,1689833241906,b2090cf9761ef60aa06e4fab97679bd43dfa5e5df073701ead5879d7c68f1ec5
-static/css/main.45332771.css,1690486789177,af661dc175f1413e9f00f4a7357d25a7d82ddd00998023e889f98ebde64feb4e
-static/js/787.ded37e57.chunk.js.map,1690486789177,3efb485ada2ccbe522ef039d26b6556b5fba38af3512a96289fc73ca65ca0629
-static/css/main.45332771.css.map,1690486789177,46c55140793ada84855580cef27293a57b505b72acc9942c82d15ac33137de25
-static/js/787.ded37e57.chunk.js,1690486789177,63f8c07042d5b1117140a3f5f625af255e203f37ca8ec1072d663b3306b71efa
-static/js/main.5c0df9c2.js.LICENSE.txt,1690486789177,a6e209577de7530c0f883c6cf97166a4c3ba703bc1ce8467bb497ac465905c2c
-favicon.ico,1689833241906,eae62e993eb980ec8a25058c39d5a51feab118bd2100c4deebb2a9c158ec11f9
-logo512.png,1689833241906,ee7e2f3fdb8209c4b6fd7bef6ba50d1b9dba30a25bb5c3126df057e1cb6f5331
-logo192.png,1689833241906,3ee59515172ee198f3be375979df15ac5345183e656720a381b8872b2a39dc8b
-static/media/playfair.27705d9c43e8c36b1abd.ttf,1690486789177,40efd44d961f127daf96f457aff7827046df2d491f19986e0e927457dcba3b6f
-static/media/pic.47b011871ef89e88b3ed.jpg,1690486789177,396cec574a6128b9bea307bd72fc6c38e454aa522562501847b9ef0b388f2c4a
-static/js/main.5c0df9c2.js,1690486789177,64b07aa3be29ba512520b35499fce5a94ccb54c177da486407a4e0bf14d6e00d
-static/js/main.5c0df9c2.js.map,1690486789177,343e6921b46869e4872094de1ce8127dcc17fd2f8a45daa343c49b5f9e974110
