+++ package.json
@@ -19,7 +19,6 @@
     "react-icons": "^4.9.0",
     "react-otp-input": "^3.0.2",
     "react-phone-number-input": "^3.2.23",
-    "react-responsive": "^9.0.2",
     "react-router-dom": "^6.11.2",
     "react-scripts": "5.0.1",
     "react-toastify": "^9.1.3",
