+++ src/components/login.css
@@ -53,19 +53,3 @@
     background-color: #406462 !important;
     color: #ffffff !important;
   }
-
-  @media (max-width: 600px) {
-    .mobileCard {
-      background: url('../../images/pic.jpg')  center center fixed;
-      background-size: cover;
-    }
-  }
-  @media (max-width: 900px) {
-    .mobileCard {
-      position: absolute;
-      top: 50%;
-      left: 50%;
-      transform: translate(-50%, -50%) scale(0.70);
-      width: 80%;
-    }
-  }
