+++ src/components/privateRoute.js
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react';
 import { Navigate, useLocation } from 'react-router-dom';
-import { useAuth } from '../../context/authContext';
+import { useAuth } from '../context/authContext';
 
 const PrivateRoute = ({ component}) => {
     const { currentUser } = useAuth();
