+++ src/components/profile.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+// Your components and additional imports go here
+
+function Profile() {
+  return (
+    <div>
+      <h1>Profile</h1>
+    </div>
+  );
+}
+export default Profile;
